relocate/rename some stuff
